Update idasen to 0.12; Bump to 2.5.2 (#25)
https://github.com/newAM/idasen/pull/409 got merged, so the version can
be bumped to 0.12.0 again. This improves the height changes quite a bit,
it will not overshoot or undershoot anymore, but always stop exactly at
the requested height (same as with the IKEA app). Also speed got added,
so once the version is updated, this can also be added as a sensor to
HA.